+++ jgclark.Dashboard/src/react/components/Header/SearchBar.jsx
@@ -1,19 +1,45 @@
+// @flow
+//--------------------------------------------------------------------------
+// Dashboard React component to show a search bar in the Header.
+// Last updated for v2.2.0.a4
+//--------------------------------------------------------------------------
+
 import React, { useEffect, useRef, useState } from 'react'
 import './SearchBar.css'
 
-const SearchBar = ({ onSearch }) => {
+//----------------------------------------------------------------------
+
+type Props = {
+  onSearch: (query: string) => void,
+}
+
+const SearchBar = ({ onSearch }: Props) => {
+  //----------------------------------------------------------------------
+  //Refs
+  //----------------------------------------------------------------------
+
+  const searchContainerRef = useRef <? HTMLDivElement > (null)
+
+//----------------------------------------------------------------------
+// State
+//----------------------------------------------------------------------
+
   const [isActive, setIsActive] = useState(false)
   const [query, setQuery] = useState('')
-  const searchContainerRef = useRef(null)
+
+  //----------------------------------------------------------------------
+  // Functions
+  //----------------------------------------------------------------------
+
   const handleIconClick = () => {
     setIsActive(!isActive)
   }
 
-  const handleInputChange = (event) => {
+  const handleInputChange = (event: SyntheticInputEvent<HTMLInputElement>) => {
     setQuery(event.target.value)
   }
 
-  const handleKeyPress = (event) => {
+  const handleKeyPress = (event: KeyboardEvent) => {
     if (event.key === 'Enter' && onSearch) {
       console.log(`SearchBar: handleKeyPress: Enter key pressed, with query term currently '${query}'`) // OK here
       onSearch(query)
@@ -21,11 +47,23 @@ const SearchBar = ({ onSearch }) => {
     }
   }
 
-  const handleClickOutside = (event) => {
-    if (searchContainerRef.current && !searchContainerRef.current.contains(event.target)) {
+  const handleClickOutside = (event: MouseEvent) => {
+    const target = event.target
+    if (searchContainerRef.current && target instanceof Node && !searchContainerRef.current.contains(target)) {
       setIsActive(false)
     }
   }
+
+  //----------------------------------------------------------------------
+  // Effects
+  //----------------------------------------------------------------------
+
+  // Force the window to be focused on load so that we can capture clicks on hover
+  useEffect(() => {
+    if (searchContainerRef.current) {
+      searchContainerRef.current.focus()
+    }
+  }, [])
 
   useEffect(() => {
     document.addEventListener('mousedown', handleClickOutside)